requirements.txt
azure-cognitiveservices-language-luis
azure-functions
black
pandas
pytest